Move export tasks into exports module
