+++ corehq/apps/export/tasks.py
@@ -1,24 +1,36 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
+from datetime import datetime, timedelta
 import logging
-from celery.task import task
+from celery.schedules import crontab
+from celery.task import task, periodic_task
+from django.conf import settings
 
 from corehq.apps.data_dictionary.util import add_properties_to_data_dictionary
-from corehq.apps.export.export import get_export_file, rebuild_export, should_rebuild_export
-from corehq.apps.export.dbaccessors import get_case_inferred_schema, get_properly_wrapped_export_instance
-from corehq.apps.export.system_properties import MAIN_CASE_TABLE_PROPERTIES
-from corehq.apps.export.models.new import EmailExportWhenDoneRequest
+from corehq.apps.reports.models import HQGroupExportConfiguration
 from corehq.apps.users.models import CouchUser
+from corehq.blobs import get_blob_db
+from corehq.dbaccessors.couchapps.all_docs import get_doc_ids_by_class
 from corehq.util.datadog.gauges import datadog_track_errors
 from corehq.util.decorators import serial_task
 from corehq.util.files import safe_filename_header, TransientTempfile
 from corehq.util.quickcache import quickcache
-from corehq.blobs import get_blob_db
+from couchexport.groupexports import export_for_group
 from couchexport.models import Format
 from dimagi.utils.couch import CriticalSection
 from soil.util import expose_blob_download, process_email_request
-from six.moves import filter
 
+from .const import SAVED_EXPORTS_QUEUE
+from .dbaccessors import (
+    get_case_inferred_schema,
+    get_properly_wrapped_export_instance,
+    get_all_daily_saved_export_instance_ids,
+)
+from .export import get_export_file, rebuild_export, should_rebuild_export
+from .models.new import EmailExportWhenDoneRequest
+from .system_properties import MAIN_CASE_TABLE_PROPERTIES
+
+from six.moves import filter
 
 logger = logging.getLogger('export_migration')
 
@@ -89,6 +101,33 @@ def manually_rebuild_export_task(export_instance_id, last_access_cutoff=None, fi
 @serial_task('{domain}-{case_type}', queue='background_queue')
 def add_inferred_export_properties(sender, domain, case_type, properties):
     _cached_add_inferred_export_properties(sender, domain, case_type, properties)
+
+
+@task(queue=SAVED_EXPORTS_QUEUE, ignore_result=True)
+def export_for_group_async(group_config_id):
+    # exclude exports not accessed within the last 7 days
+    last_access_cutoff = datetime.utcnow() - timedelta(days=settings.SAVED_EXPORT_ACCESS_CUTOFF)
+    group_config = HQGroupExportConfiguration.get(group_config_id)
+    export_for_group(group_config, last_access_cutoff=last_access_cutoff)
+
+
+@periodic_task(run_every=crontab(hour="23", minute="59", day_of_week="*"),
+               queue=getattr(settings, 'CELERY_PERIODIC_QUEUE', 'celery'))
+def saved_exports():
+    for group_config_id in get_doc_ids_by_class(HQGroupExportConfiguration):
+        export_for_group_async.delay(group_config_id)
+
+    for daily_saved_export_id in get_all_daily_saved_export_instance_ids():
+        last_access_cutoff = datetime.utcnow() - timedelta(days=settings.SAVED_EXPORT_ACCESS_CUTOFF)
+        rebuild_export_task.apply_async(
+            args=[
+                daily_saved_export_id, last_access_cutoff
+            ],
+            # Normally the rebuild_export_task uses the background queue,
+            # however we want to override it to use its own queue so that it does
+            # not disrupt other actions.
+            queue=SAVED_EXPORTS_QUEUE,
+        )
 
 
 @quickcache(['sender', 'domain', 'case_type', 'properties'], timeout=60 * 60)
