+++ corehq/apps/reports/tasks.py
@@ -19,14 +19,12 @@ from celery.task import task
 from celery.utils.log import get_task_logger
 
 from casexml.apps.case.xform import extract_case_blocks
-from corehq.apps.export.dbaccessors import get_all_daily_saved_export_instance_ids
 from corehq.apps.export.const import SAVED_EXPORTS_QUEUE
 from corehq.apps.reports.util import send_report_download_email
-from corehq.dbaccessors.couchapps.all_docs import get_doc_ids_by_class
 from corehq.form_processor.interfaces.dbaccessors import FormAccessors
 from corehq.util.dates import iso_string_to_datetime
 from couchexport.files import Temp
-from couchexport.groupexports import export_for_group, rebuild_export
+from couchexport.groupexports import rebuild_export
 from couchexport.tasks import cache_file_to_be_served
 from couchforms.analytics import app_has_been_submitted_to_in_last_30_days
 from dimagi.utils.couch.cache.cache_core import get_redis_client
@@ -54,7 +52,6 @@ from .analytics.esaccessors import (
 from .export import save_metadata_export_to_tempfile
 from .models import (
     FormExportSchema,
-    HQGroupExportConfiguration,
     ReportNotification,
     UnsupportedScheduledReportError,
 )
@@ -143,40 +140,6 @@ def weekly_reports():
 def monthly_reports():
     for report_id in get_scheduled_report_ids('monthly'):
         send_delayed_report(report_id)
-
-
-@periodic_task(run_every=crontab(hour="23", minute="59", day_of_week="*"), queue=getattr(settings, 'CELERY_PERIODIC_QUEUE', 'celery'))
-def saved_exports():
-    for group_config_id in get_doc_ids_by_class(HQGroupExportConfiguration):
-        export_for_group_async.delay(group_config_id)
-
-    for daily_saved_export_id in get_all_daily_saved_export_instance_ids():
-        from corehq.apps.export.tasks import rebuild_export_task
-        last_access_cutoff = datetime.utcnow() - timedelta(days=settings.SAVED_EXPORT_ACCESS_CUTOFF)
-        rebuild_export_task.apply_async(
-            args=[
-                daily_saved_export_id, last_access_cutoff
-            ],
-            # Normally the rebuild_export_task uses the background queue,
-            # however we want to override it to use its own queue so that it does
-            # not disrupt other actions.
-            queue=SAVED_EXPORTS_QUEUE,
-        )
-
-
-@task(queue='background_queue', ignore_result=True)
-def rebuild_export_task(groupexport_id, index, last_access_cutoff=None, filter=None):
-    group_config = HQGroupExportConfiguration.get(groupexport_id)
-    config, schema = group_config.all_exports[index]
-    rebuild_export(config, schema, last_access_cutoff, filter=filter)
-
-
-@task(queue=SAVED_EXPORTS_QUEUE, ignore_result=True)
-def export_for_group_async(group_config_id):
-    # exclude exports not accessed within the last 7 days
-    last_access_cutoff = datetime.utcnow() - timedelta(days=settings.SAVED_EXPORT_ACCESS_CUTOFF)
-    group_config = HQGroupExportConfiguration.get(group_config_id)
-    export_for_group(group_config, last_access_cutoff=last_access_cutoff)
 
 
 @task(queue=SAVED_EXPORTS_QUEUE, ignore_result=True)
